+++ Models/ResultGrpProductModel.cs
@@ -5,9 +5,9 @@
 		public string ProductID { get; set; }
 		public string ProductName { get; set; }
         public string SectionName { get; set; }
-        public int STD { get; set; }
-		public int Actual { get; set; }
-		public int Diff { get; set; }
+        public double STD { get; set; }
+		public double Actual { get; set; }
+		public double Diff { get; set; }
 
 	}
 }
